feat(mrc): drawer composable
ref: #MANAGER-19661
