--- packages/manager-react-components/src/components/drawer/Drawer.component.tsx
@@ -1,14 +1,0 @@
-import DrawerBackdrop from './DrawerBackdrop.component';
-import { DrawerBase, DrawerBaseProps } from './DrawerBase.component';
-import './translations';
-
-export type DrawerProps = Omit<DrawerBaseProps, 'className'>;
-
-export const Drawer = (props: DrawerProps) => {
-  return (
-    <div id="mrc-drawer" className="relative">
-      <DrawerBase {...props} />
-      {props.isOpen && <DrawerBackdrop onClick={props.onDismiss} />}
-    </div>
-  );
-};
