+++ packages/manager-react-components/src/components/drawer/drawer-backdrop/DrawerBackdrop.props.tsx
@@ -1,0 +1,3 @@
+export type DrawerBackdropProps = {
+  onClick: () => void;
+};
