+++ packages/manager-react-components/src/components/drawer/drawer-base/DrawerBase.props.tsx
@@ -1,0 +1,12 @@
+import { PropsWithChildren } from 'react';
+
+export type DrawerBaseProps = PropsWithChildren & {
+  /** Open/close the drawer (default to true) */
+  isOpen?: boolean;
+  /** Callback function to be called on close of the Drawer */
+  onDismiss: () => void;
+  /** Show a loader instead of the drawer content  */
+  isLoading?: boolean;
+  /** Class name for the drawer */
+  className?: string;
+};
