+++ packages/manager-react-components/src/components/drawer/drawer-content/DrawerContent.component.tsx
@@ -1,0 +1,9 @@
+import { PropsWithChildren } from 'react';
+
+export const DrawerContent = ({ children }: PropsWithChildren) => {
+  return (
+    <section className="px-6 flex-1 overflow-y-auto outline-none">
+      {children}
+    </section>
+  );
+};
