+++ packages/manager-react-components/src/components/drawer/drawer-footer/DrawerFooter.props.tsx
@@ -1,0 +1,18 @@
+export type DrawerFooterProps = {
+  /** Primary Button Label */
+  primaryButtonLabel?: string;
+  /** Primary Button Loading State */
+  isPrimaryButtonLoading?: boolean;
+  /** Primary Button Disabled State */
+  isPrimaryButtonDisabled?: boolean;
+  /** Primary Button Callback */
+  onPrimaryButtonClick?: () => void;
+  /** Secondary Button Label */
+  secondaryButtonLabel?: string;
+  /** Secondary Button Disabled State */
+  isSecondaryButtonDisabled?: boolean;
+  /** Secondary Button Loading State */
+  isSecondaryButtonLoading?: boolean;
+  /** Secondary Button Callback */
+  onSecondaryButtonClick?: () => void;
+};
