+++ packages/manager-react-components/src/components/drawer/drawer-handle/DrawerHandle.props.tsx
@@ -1,0 +1,6 @@
+import { DrawerCollapseState } from '../drawer.types';
+
+export type DrawerHandleProps = {
+  onClick: () => void;
+  collapseState: DrawerCollapseState;
+};
