+++ packages/manager-react-components/src/components/drawer/drawer-header/DrawerHeader.props.tsx
@@ -1,0 +1,3 @@
+export type DrawerHeaderProps = {
+  title: string;
+};
