+++ packages/manager-react-components/src/components/drawer/drawer-header/__tests__/DrawerHeader.test.tsx
@@ -1,0 +1,9 @@
+import { assertTextVisibility } from '@ovh-ux/manager-core-test-utils';
+import { render } from '@testing-library/react';
+import { DrawerHeader } from '../DrawerHeader.component';
+
+it('should display the title', async () => {
+  render(<DrawerHeader title="Drawer header title" />);
+
+  await assertTextVisibility('Drawer header title');
+});
