+++ packages/manager-react-components/src/components/drawer/drawer-root-collapsible/DrawerRootCollapsible.props.tsx
@@ -1,0 +1,3 @@
+import { DrawerBaseProps } from '../drawer-base/DrawerBase.props';
+
+export type DrawerRootCollapsibleProps = Omit<DrawerBaseProps, 'className'>;
