+++ packages/manager-react-components/src/components/drawer/drawer-root/DrawerRoot.props.tsx
@@ -1,0 +1,3 @@
+import { DrawerBaseProps } from '../drawer-base/DrawerBase.props';
+
+export type DrawerRootProps = Omit<DrawerBaseProps, 'className'>;
