+++ packages/manager-react-components/src/components/drawer/index.ts
@@ -1,0 +1,1 @@
+export * as Drawer from './namespace';
