+++ packages/manager-react-components/src/components/drawer/namespace.tsx
@@ -1,0 +1,9 @@
+export { DrawerRoot as Root } from './drawer-root/DrawerRoot.component';
+export { DrawerRootCollapsible as RootCollapsible } from './drawer-root-collapsible/DrawerRootCollapsible.component';
+export { DrawerContent as Content } from './drawer-content/DrawerContent.component';
+export { DrawerHeader as Header } from './drawer-header/DrawerHeader.component';
+export { DrawerFooter as Footer } from './drawer-footer/DrawerFooter.component';
+export type { DrawerFooterProps as FooterProps } from './drawer-footer/DrawerFooter.props';
+export type { DrawerRootProps as RootProps } from './drawer-root/DrawerRoot.props';
+export type { DrawerRootCollapsibleProps as RootCollapsibleProps } from './drawer-root-collapsible/DrawerRootCollapsible.props';
+export type { DrawerHeaderProps as HeaderProps } from './drawer-header/DrawerHeader.props';
