+++ packages/manager-react-components/src/components/index.ts
@@ -17,8 +17,7 @@ export * from './datagrid/useDatagrid';
 export * from './datagrid/useDatagridSearchParams';
 export * from './datagrid/clipboard-cell.component';
 
-export * from './drawer/Drawer.component';
-export * from './drawer/DrawerCollapsible.component';
+export * from './drawer';
 
 export * from './guides-header';
